implemented new node lib refactor
